Refactor doctor consultation client for faster builds
